[Refactoring] pages 폴더로 페이지를 구성하는 컴포넌트들 이동
