+++ react-app/src/App.js
@@ -1,149 +1,9 @@
-import React, { useState } from 'react';
+import React from 'react';
 import 'react-calendar/dist/Calendar.css';
-import { Link, Routes, Route } from 'react-router-dom';
-import WorkTagTable from './work-tag-table';
-import moment from 'moment';
-import Calendar from 'react-calendar';
-import Modal from './components/Modal';
-
-function WorkSetting() {
-  return (
-    <div>
-      <div>
-        <label>급여시작일</label>
-        <input type="text" placeholder="입력"></input>
-        <br />
-        <label>급여종료일</label>
-        <input type="text" placeholder="입력"></input>
-      </div>
-      <label>기본 단가</label>
-      <input type="text" placeholder="입력"></input>
-      <div>
-        <h3>나의 근로 태그 목록</h3>
-        <WorkTagTable />
-      </div>
-      <button>추가</button>
-    </div>
-  );
-}
-
-function WorkingPeriod(props) {
-  return (
-    <div>
-      {moment(props.value).format('YYYY년 MM월 DD일')}부터{' '}
-      {moment(props.value).add(1, 'months').format('YYYY년 MM월 DD일')}까지{' '}
-      <span>상세 내역</span>
-    </div>
-  );
-}
-
-function WorkInfo(props) {
-  const workInfo = props.data.map((info) => {
-    return (
-      <span key={info.id}>
-        {info.content}
-        <br />
-      </span>
-    );
-  });
-  return <div>{workInfo}</div>;
-}
-function Main() {
-  const [loginModal, setLoginModal] = useState(true);
-  const [signUpModal, setSignUpModal] = useState(true);
-  return (
-    <div>
-      <button onClick={() => setSignUpModal(!signUpModal)}>SignUp</button>
-      {!signUpModal && (
-        <Modal
-          closeModal={() => {
-            setSignUpModal(!signUpModal);
-          }}
-        >
-          <SignUp signUpModal={signUpModal} setSignUpModal={setSignUpModal} />
-        </Modal>
-      )}
-      <button onClick={() => setLoginModal(!loginModal)}>Login</button>
-      {!loginModal && (
-        <Modal
-          closeModal={() => {
-            setLoginModal(!loginModal);
-          }}
-        >
-          <Login loginModal={loginModal} setLoginModal={setLoginModal} />
-        </Modal>
-      )}
-    </div>
-  );
-}
-function CalendarInfo() {
-  const [value, onChange] = useState(new Date());
-
-  const elements = [
-    { id: 1, content: '총 공수' },
-    { id: 2, content: '총 급여' },
-    { id: 3, content: '총 근무일' },
-  ];
-  return (
-    <div>
-      <Calendar onChange={onChange} value={value}></Calendar>
-      <WorkingPeriod value={value} />
-      <WorkInfo data={elements} />
-      <Link to="/workSetting">
-        <button>설정</button>
-      </Link>
-    </div>
-  );
-}
-
-function Login(props) {
-  function loginCheck() {
-    // db에 일치하는 이메일 비번 있는지 확인 후
-    props.setLoginModal(!props.loginModal); // 일치하면 모달 창 닫고 캘린더로 이동
-    // 일치하는 계정이 없으면 alert창 띄우기
-  }
-  return (
-    <div>
-      <div>
-        <label>이메일</label>
-        <input type="email" placeholder="이메일을 입력하세요."></input>
-      </div>
-      <div>
-        <label>비밀번호</label>
-        <input type="password" placeholder="비밀번호를 입력하세요."></input>
-      </div>
-      <div>
-        <Link to="/findPassword">비밀번호 찾기</Link>
-      </div>
-      <button onClick={loginCheck}>로그인</button>
-    </div>
-  );
-}
-
-function SignUp(props) {
-  function signUpCheck() {
-    // 일치하는 계정 확인
-    // 없으면 가입 후 자동 로그인
-    props.setSignUp(!props.signUp); // 모달 창 닫기
-  }
-  return (
-    <div>
-      <div>
-        <label>이메일</label>
-        <input type="email" placeholder="이메일을 입력하세요"></input>
-      </div>
-      <div>
-        <label>닉네임</label>
-        <input type="email" placeholder="닉네임을 원하시면 입력하세요"></input>
-      </div>
-      <div>
-        <label>비밀번호</label>
-        <input type="password" placeholder="비밀번호를 입력하세요"></input>
-      </div>
-      <button onClick={signUpCheck}>가입하기</button>
-    </div>
-  );
-}
+import { Routes, Route } from 'react-router-dom';
+import WorkSetting from './pages/WorkSetting';
+import CalendarInfo from './pages/CalendarInfo';
+import Main from './pages/Main';
 
 function App() {
   return (
